LayoutTests/platform/chromium/fast/events/touch/gesture/touch-gesture-scroll-div-expected.txt

This tests gesture event scrolling of an overflow div.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
PASS movingdiv.scrollTop is 0
PASS movingdiv.scrollLeft is 90
PASS wheelEventsOccurred is 0
second gesture
PASS movingdiv.scrollTop is 95
PASS movingdiv.scrollLeft is 90
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDivElement]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
